Add in scan option so that you can scan files without changing them

+++ s3ObjectValidator.py
@@ -1,38 +1,45 @@
 import re, urllib.parse, random
 from pathlib import Path
 
-def makeFullStructureSanitize(rootdir,history,filesBool=False,dirsBool=False, verboseBool = False):
+def makeFullStructureSanitize(rootdir,history,filesBool=False,dirsBool=False, DesructiveBool = True, verboseBool = False):
 
     for path in Path(rootdir).iterdir():
         if path.is_symlink():
             if verboseBool:
-                print("unlinking symlink")
-                print(path)
-            path.unlink()
+                print(f"{path} is a symlink, symlinks will be removed as they are not supported in S3")
+            print("unlinking symlink")
+            print(path)
+            if DesructiveBool:
+                path.unlink()
         if path.is_file() and filesBool:
             # history['before']['files'].append(path)
             if needsSanitizing(path):
                 if verboseBool:
-                    print("changing file")
-                    print(path)
-                path = sanitize(path)
+                    print(f"{path} is a file")
+                print(f"{path} needs to be sanitized")
+                print(path)
+                if DesructiveBool:
+                    path = sanitize(path)
                 # history['after']['files'].append(path)
         elif path.is_dir() and not path.is_symlink():
             if dirsBool:
                 # history['before']['dirs'].append(path)
                 if needsSanitizing(path):
                     if verboseBool:
-                        print("changing dir")
-                        print(path)
-                    path = sanitize(path)
+                        print(f"{path} is a directory")
+                    print(f"{path} needs to be sanitized")
+                    print(path)
+                    if DesructiveBool:
+                        path = sanitize(path)
                     # history['after']['dirs'].append(path)
-            makeFullStructureSanitize(path,history,filesBool,dirsBool)
+            print(f"Changing directory to {path}")
+            makeFullStructureSanitize(path,history,filesBool,dirsBool,DesructiveBool,verboseBool)
 
 
-def sanitizeStructure(rootdir,filesBool=False,dirsBool=False, verboseBool = False):
+def sanitizeStructure(rootdir,filesBool=False,dirsBool=False, DesructiveBool = True,verboseBool = False):
     history = {"before": {"files" : [] , "dirs" : [] }, "after": {"files" : [] , "dirs" : [] }}
 
-    makeFullStructureSanitize(rootdir,history,filesBool,dirsBool,verboseBool)
+    makeFullStructureSanitize(rootdir,history,filesBool,dirsBool,DesructiveBool,verboseBool)
 
     if filesBool:
         return { "before" : history['before']['files'], "after" : history['after']['files']}
